Implement Cart

+++ src/main/java/org/example/sotre/controller/CartController.java
@@ -2,7 +2,9 @@ package org.example.sotre.controller;
 
 import jakarta.validation.constraints.Min;
 import jakarta.validation.constraints.NotBlank;
+import org.example.sotre.controller.exceptions.NoStockException;
 import org.example.sotre.model.Cart;
+import org.example.sotre.response.CartResponse;
 import org.example.sotre.service.CartService;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.access.prepost.PreAuthorize;
@@ -25,58 +27,99 @@ public class CartController {
     }
     @PreAuthorize("hasAnyRole('ADMIN','USER')")
 
-    @GetMapping("/{email}/latest")
-    public ResponseEntity<Cart> getLatestCart(@PathVariable @NotBlank String email) {
-        String user = (String) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
-        System.out.println(user+" PARSA");
+    @GetMapping("/latest")
+    public ResponseEntity<CartResponse> getLatestCart() {
+        String email = (String) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        if(email == null) {
+            throw new RuntimeException("Invalid User");
+        }
         Cart cart = cartService.getLatestCart(email);
         if (cart == null) {
             return ResponseEntity.notFound().build();
         }
 
-        return ResponseEntity.ok(cart);
+        return ResponseEntity.ok(CartResponse.from(cart));
     }
     @PreAuthorize("hasAnyRole('ADMIN','USER')")
 
-    @GetMapping("/{email}/carts")
-    public ResponseEntity<List<Cart>> getAllUserCarts(@PathVariable @NotBlank String email) {
+    @GetMapping("/carts")
+    public ResponseEntity<List<CartResponse>> getAllUserCarts() {
+        String email = (String) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        if(email == null) {
+            throw new RuntimeException("Invalid User");
+        }
         List<Cart> cart = cartService.getCarts(email);
         if (cart == null) {
             return ResponseEntity.notFound().build();
         }
-        return ResponseEntity.ok(cart);
+        return ResponseEntity.ok(CartResponse.from(cart));
     }
     @PreAuthorize("hasRole('ADMIN')")
     @GetMapping("/")
-    public ResponseEntity<List<Cart>> getAllCarts() {
+    public ResponseEntity<List<CartResponse>> getAllCarts() {
         List<Cart> cart = cartService.getCarts();
         if (cart == null) {
             return ResponseEntity.notFound().build();
         }
-        return ResponseEntity.ok(cart);
+        return ResponseEntity.ok(CartResponse.from(cart));
     }
     @PreAuthorize("hasAnyRole('ADMIN','USER')")
 
-    @PostMapping("/{email}/add/{productId}")
-    public ResponseEntity<Cart> addProductToCart(
-            @PathVariable @NotBlank String email,
+    @PostMapping("/add/{productId}")
+    public ResponseEntity<CartResponse> addProductToCart(
             @PathVariable @Min(1) int productId) {
+        String email = (String) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        if(email == null) {
+            throw new RuntimeException("Invalid User");
+        }
+        Cart updatedCart = null;
 
-        Cart updatedCart = cartService.addProduct(email, productId);
-        return ResponseEntity.ok(updatedCart);
+     updatedCart = cartService.addProduct(email, productId);
+
+        return ResponseEntity.ok(CartResponse.from(updatedCart));
     }
     @PreAuthorize("hasAnyRole('ADMIN','USER')")
 
-    @DeleteMapping("/{email}/remove/{productId}")
+    @DeleteMapping("/remove/{productId}")
     public ResponseEntity<String> removeProductFromCart(
-            @PathVariable @NotBlank String email,
             @PathVariable @Min(1) int productId) {
-
+        String email = (String) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        if(email == null) {
+            throw new RuntimeException("Invalid User");
+        }
         boolean removed = cartService.removeProduct(email, productId);
         if (removed) {
             return ResponseEntity.ok("Product removed successfully.");
         } else {
             return ResponseEntity.badRequest().body("Product not found in cart.");
         }
+    }
+
+
+    @PreAuthorize("hasRole('ADMIN')")
+
+    @GetMapping("/carts/{cartId}")
+    public ResponseEntity<CartResponse> getAllUserCarts(
+            @PathVariable @Min(1) int cartId) {
+        String email = (String) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        if(email == null) {
+            throw new RuntimeException("Invalid User");
+        }
+        Cart cart = cartService.getCart(cartId);
+        if (cart == null) {
+            return ResponseEntity.notFound().build();
+        }
+        return ResponseEntity.ok(CartResponse.from(cart));
+    }
+    @PreAuthorize("hasRole('ADMIN')")
+    @GetMapping("/carts/{userId}/carts")
+    public ResponseEntity<List<CartResponse>> getAllUserCarts(
+            @PathVariable @NotBlank String userId) {
+
+        List<Cart> carts = cartService.getCarts(userId);
+        if (carts == null) {
+            return ResponseEntity.notFound().build();
+        }
+        return ResponseEntity.ok(CartResponse.from(carts));
     }
 }
